refactor(Number baseball game): Move the position of the class.
- 클래스들의 위치 이동
