+++ src/main/java/baseball/number/NumberConstant.java
@@ -1,8 +1,8 @@
 package baseball.number;
 
 public class NumberConstant {
-    static final int MINIMUM_NUMBER = 1;
-    static final int MAXIMUM_NUMBER = 9;
-    static final int INITIAL_INDEX = 0;
-    static final int COUNT_OF_NUMBERS = 3;
+    public static final int MINIMUM_NUMBER = 1;
+    public static final int MAXIMUM_NUMBER = 9;
+    public static final int INITIAL_INDEX = 0;
+    public static final int COUNT_OF_NUMBERS = 3;
 }
