+++ src/main/java/baseball/number/strategy/ExtractRandomNumberByShuffleList.java
@@ -1,4 +1,4 @@
-package baseball.number;
+package baseball.number.strategy;
 
 import java.util.ArrayList;
 import java.util.Collections;
