+++ src/test/java/baseball/number/matcher/NumberMatcherTest.java
@@ -1,5 +1,7 @@
-package baseball.number;
+package baseball.number.matcher;
 
+import baseball.number.InputNumbers;
+import baseball.number.TargetNumbers;
 import baseball.score.Score;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
