+++ src/test/java/baseball/number/strategy/ExtractRandomNumberByPickNumberTest.java
@@ -1,4 +1,4 @@
-package baseball.number;
+package baseball.number.strategy;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
